controller-rpi/pyproject.toml
[tool.poetry]
name = "rattechan-controller-rpi"
version = "0.1.0"
description = ""
authors = ["Valentin Safronov <svdouble@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
paho-mqtt = "^2.1.0"
numpy = "^2.1.2"
pydantic = "^2.10.1"
pydantic-settings = "^2.6.1"
rich = "13.9.4"
gpiozero = "2.0.1"
lgpio = "^0.2.2.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.8.0"
pre-commit = "^4.0.1"

[tool.ruff]
line-length = 88
lint.select = ["E", "F", "W", "C", "N", "B"]
# Disallow all relative imports.
lint.flake8-tidy-imports.ban-relative-imports = "all"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
